fold MergeResourceRequirements into trainingruntime.go
